Muscop update:
Muscop v2.0: In this version Muscop is able to handle the loss of
leader(s) and/or slave(s). This is done by keeping track if all the UAVs
are still alive (by listing to the messages they already send) and if a
UAV is not alive act accordingly. Due to this changes the protocol is
now also able to handle a swarm split-up. In order to test this a new
flightformation was created "split-up".